requirements.txt
appdirs==1.4.0
beautifulsoup4==4.5.3
Django==1.10.5
django-modelcluster==2.0
django-taggit==0.18.3
django-treebeard==4.1.0
djangorestframework==3.5.4
html5lib==0.999999999
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyparsing==2.1.10
pytz==2016.10
requests==2.13.0
six==1.10.0
suds-jurko==0.6
Unidecode==0.4.20
wagtail==1.8.1
webencodings==0.5
Willow==0.4
